finishing printf TODO refacto + norme

+++ printf_header.h
@@ -6,7 +6,7 @@
 /*   By: brguicho <brguicho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/10 10:02:26 by brguicho          #+#    #+#             */
-/*   Updated: 2023/11/13 14:25:02 by brguicho         ###   ########.fr       */
+/*   Updated: 2023/11/13 15:08:58 by brguicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,5 +18,5 @@
 
 int				ft_printf(const char *, ...);
 int				ft_putstrlen(char *str);
-long long int	ft_putnbr_basel(int nbr, char *base);
+int	ft_putnbr_basel( long long int nbr, char *base);
 #endif
